fix: Refactor data fetching to fix blank page issue
The application was displaying a blank page because it was fetching data from a public Google Sheet URL directly from the frontend. This method is unreliable and was likely failing in the production environment.

This commit refactors the data fetching logic to use a more robust, backend-driven approach:

1.  **New API Endpoint:** A new API endpoint `/api/documents` has been created on the Express server. This endpoint uses the authenticated `googleapis` client to fetch data from the Google Sheet, ensuring a more reliable connection.
2.  **Frontend Refactoring:** The `documentStore.ts` on the frontend has been updated to call the new `/api/documents` endpoint instead of the public URL. This also includes updating the data types to handle the JSON response correctly.
3.  **Removed Unused Code:** The `googleSheetsService.ts` file, which contained the old, unreliable fetching logic, has been removed.

**Note:** Due to persistent timeouts in the sandbox environment, I was unable to run the application locally to verify these changes. However, the implemented solution is based on the existing application patterns and addresses the identified root cause of the issue.
