+++ package.json
@@ -102,6 +102,7 @@
     "typescript": "^5.5.3",
     "vaul": "^0.9.3",
     "vite": "^6.2.2",
-    "vitest": "^3.1.4"
+    "vitest": "^3.1.4",
+    "npm-run-all": "^4.1.5"
   }
 }
